Split compiler driver into separate components (#2453)
**Context:** This PR divides the compiler driver into multiple
components, allowing for better isolation of the corresponding
interactions between them. This is a precursor to updating the target
generation layer for multiple output formats.

**Description of the Change:** 
* Isolated the CompilerDriver implementation to use multiple
headers/translation units
* Added explicit namespace scoping to make clear delineations for source
of truth (fixes issues with clangd)
* Added more precise timing control of the instrumentation functionality
using `std::chrono` defined ratios (ns, us, ms all supported).
* Defined a `main` driver method that ties all the surrounded
functionality into a single location.

**Benefits:**
* Makes isolation of the target definition and generation calls easier
* Static analysis helps to identify the paths to and through MLIR, LLVM,
and our implementations
* Access to higher-precision timer divisions for the instrumentation
layer via templated arguments

**Possible Drawbacks:**
* Potential change in compilation times due to the restructured driver
(LTO may be warranted) --- UPDATE: LTO did not improve times.

**Related GitHub Issues:**

---------
