+++ mlir/include/Driver/CompilerDriver.h
@@ -19,28 +19,56 @@
 #include <unordered_map>
 #include <vector>
 
-#include "llvm/ADT/SmallVector.h"
+#include "llvm/ADT/StringRef.h"     // llvm::StringRef
+#include "llvm/IR/Module.h"         // llvm::Module
+#include "llvm/Support/SourceMgr.h" // llvm::SMDiagnostic
 #include "llvm/Support/raw_ostream.h"
 
-#include "mlir/IR/MLIRContext.h"
-#include "mlir/Support/LogicalResult.h"
+#include "mlir/IR/BuiltinOps.h"         // mlir::ModuleOp
+#include "mlir/IR/DialectRegistry.h"    // mlir::DialectRegistry
+#include "mlir/IR/MLIRContext.h"        // mlir::MLIRContext
+#include "mlir/IR/OwningOpRef.h"        // mlir::OwningOpRef
+#include "mlir/Pass/PassManager.h"      // mlir::PassManager
+#include "mlir/Support/LogicalResult.h" // llvm::LogicalResult
+#include "mlir/Support/Timing.h"
 #include "mlir/Tools/mlir-opt/MlirOptMain.h"
 
 #include "Driver/Pipelines.h"
+#include "Driver/Timer.h"
 
-namespace catalyst {
-namespace driver {
+namespace catalyst::driver {
 
-/// Verbosity level
 // TODO: Adjust the number of levels according to our needs. MLIR seems to print few really
 // low-level messages, we might want to hide these.
+/**
+ * @brief Controls the verbosity level of the compiler during processing.
+ *
+ */
 enum class Verbosity { Silent = 0, Urgent = 1, Debug = 2, All = 3 };
 
+/**
+ * @brief Controls the stage for dumping the IR.
+ *
+ */
 enum SaveTemps { None, AfterPipeline, AfterPassChanged, AfterPass };
 
+/**
+ * @brief Defines the different functional stages of the driver.
+ *
+ */
 enum Action { OPT, Translate, LLC, All };
 
+/**
+ * @brief Defines the format of the input data.
+ *
+ */
 enum InputType { MLIR, LLVMIR, OTHER };
+
+/**
+ * @brief Provides enum-defined error codes for the compiler driver.
+ *
+ */
+enum class ErrorCode : int { Success = 0, Failure = 1 };
 
 /// Helper verbose reporting macro.
 #define CO_MSG(opt, level, op)                                                                     \
@@ -50,7 +78,10 @@ enum InputType { MLIR, LLVMIR, OTHER };
         }                                                                                          \
     } while (0)
 
-/// Optional parameters, for which we provide reasonable default values.
+/**
+ * @brief Optional parameters for the compiler, for which we provide reasonable default values.
+ *
+ */
 struct CompilerOptions {
     /// The textual IR (MLIR or LLVM IR)
     mlir::StringRef source;
@@ -83,15 +114,16 @@ struct CompilerOptions {
     bool shouldEmitBytecode;
 
     /// Get the destination of the object file at the end of compilation.
-    std::string getObjectFile() const
-    {
-        using path = std::filesystem::path;
-        return path(workspace.str()) / path(moduleName.str() + ".o");
-    }
+    std::string getObjectFile() const;
 };
 
+/**
+ * @brief Holds the output from the compiler, providing access to pass stage data during
+ * compilation.
+ *
+ */
 struct CompilerOutput {
-    typedef std::unordered_map<std::string, std::string> PipelineOutputs;
+    using PipelineOutputs = std::unordered_map<std::string, std::string>;
     std::string outputFilename;
     std::string outIR;
     std::string diagnosticMessages;
@@ -103,38 +135,226 @@ struct CompilerOutput {
     bool isCheckpointFound;
 
     // Gets the next pass dump file name within a pipeline folder
-    std::string nextPassDumpFilename(std::string pipelineName, std::string ext = ".mlir")
-    {
-        return std::filesystem::path(currentStage) /
-               std::filesystem::path(std::to_string(this->passCounter++) + "_" + pipelineName)
-                   .replace_extension(ext);
-    };
+    std::string nextPassDumpFilename(const std::string &pipelineName,
+                                     const std::string &ext = ".mlir");
 
     // Gets the root-level pipeline summary file name
-    std::string nextPipelineSummaryFilename(std::string pipelineName, std::string ext = ".mlir")
-    {
-        return std::filesystem::path(std::to_string(this->globalPipelineCounter) + "_After" +
-                                     pipelineName)
-            .replace_extension(ext);
-    };
+    std::string nextPipelineSummaryFilename(const std::string &pipelineName,
+                                            const std::string &ext = ".mlir");
 
     // Set the current compilation stage for organizing output files
-    void setStage(const std::string &stageName)
-    {
-        ++globalPipelineCounter;
-        currentStage = std::to_string(globalPipelineCounter) + "_" + stageName;
-        passCounter = 1;
-    }
+    void setStage(const std::string &stageName);
 };
 
-}; // namespace driver
-}; // namespace catalyst
+/**
+ * @brief Parse an MLIR module given in textual ASM representation. Any errors during parsing will
+ * be output to diagnosticStream.
+ *
+ */
+mlir::OwningOpRef<mlir::ModuleOp> parseMLIRSource(mlir::MLIRContext *ctx,
+                                                  const llvm::SourceMgr &sourceMgr);
 
-/// Entry point to the MLIR portion of the compiler.
+/**
+ * @brief Checks if the program contains gradient operations in the input MLIR module. Used to
+ * identify validity of the program with given passes.
+ *
+ * @param llvmModule
+ * @return true Gradient operations are present in the program.
+ * @return false Gradient operations are not present in the program.
+ */
+bool containsGradients(const llvm::Module &llvmModule);
+
+/**
+ * @brief Parse an LLVM module given in textual representation. Any parse errors will be output to
+ * the provided SMDiagnostic.
+ *
+ * @param context
+ * @param source
+ * @param moduleName
+ * @param err
+ * @return std::shared_ptr<llvm::Module>
+ */
+std::shared_ptr<llvm::Module> parseLLVMSource(llvm::LLVMContext &context, llvm::StringRef source,
+                                              llvm::StringRef moduleName, llvm::SMDiagnostic &err);
+
+/**
+ * @brief Register all dialects required by the Catalyst compiler to the given MLIR dialect
+ * registry.
+ *
+ * @param registry Reference to the given MLIR dialect registry. Will be modified in-place with the
+ * defined dialects.
+ */
+void registerAllCatalystDialects(mlir::DialectRegistry &registry);
+
+/**
+ * @brief Determines if the compilation stage should be executed if a checkpointStage is provided to
+ * the compiler options. This will ensure the compiler will execute only after reaching the given
+ * checkpoint.
+ *
+ * @param options Compiler configuration options.
+ * @param output Compiler output object. Modified in-place if checkpoint is not found to indicate if
+ * the current stage matches the provided stage name.
+ * @param stageName The name of the compiler stage to treat as a checkpoint.
+ * @return true Indicates the compiler should run the given stage.
+ * @return false Indicates the compiler should not run the given stage.
+ */
+bool shouldRunStage(const CompilerOptions &options, CompilerOutput &output,
+                    const std::string &stageName);
+
+/**
+ * @brief Run LLVM passes defined for asynchronous QNode execution.
+ * @details These passes are applied when making use of asynchronous QNodes when
+ * `qjit(async_qnodes=True), allowing for independent execution of multiple QNodes defined in a
+ * given program (where supported). This stage makes use of LLVM coroutine-specific passes, and
+ * provides the most benefit when having multiple QNodes in the program that can be independently
+ * executed.
+ *
+ * @param options Compiler configuration options.
+ * @param llvmModule
+ * @param output
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult runCoroLLVMPasses(const CompilerOptions &options,
+                                      std::shared_ptr<llvm::Module> llvmModule,
+                                      CompilerOutput &output);
+
+/**
+ * @brief Parse input pipelines to return a compatible compilation pipeline stage.
+ *
+ * @param catalystPipeline List of the pipeline stages in textual format
+ * @return std::vector<catalyst::driver::Pipeline> Output pipeline stage usable by the compiler.
+ */
+std::vector<catalyst::driver::Pipeline>
+parsePipelines(const llvm::cl::list<std::string> &catalystPipeline);
+
+/**
+ * @brief Find the position of a closing parenthesis character ")" for a given input string
+ * reference.
+ *
+ * @param str Input string reference.
+ * @param openParenPos Index position of the opening "(" for the input string.
+ * @return size_t Index position of the closing parenthesis character in the string. Returns `npos`
+ * if not found.
+ */
+size_t findMatchingClosingParen(llvm::StringRef str, size_t openParenPos);
+
+/**
+ * @brief Ensure the compiler options and input data type match their semantic use.
+ *
+ * @param options Catalyst compiler options.
+ * @param inType Denote the expected input type.
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult verifyInputType(const CompilerOptions &options, InputType inType);
+
+/**
+ * @brief Apply all MLIR lowering passes on the given module.
+ *
+ * @param options Compiler configuration options.
+ * @param ctx MLIR context for all operations.
+ * @param moduleOp Top-level MLIR module container.
+ * @param output Compiler output object. Modified in-place.
+ * @param timing Timer object for instrumentation of pass execution.
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult runLowering(const CompilerOptions &options, mlir::MLIRContext *ctx,
+                                mlir::ModuleOp moduleOp, CompilerOutput &output,
+                                mlir::TimingScope &timing);
+
+/**
+ * @brief Run the given compiler pipeline.
+ *
+ * @param pm
+ * @param options Compiler configuration options.
+ * @param output
+ * @param pipeline
+ * @param clHasManualPipeline
+ * @param moduleOp
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult runPipeline(mlir::PassManager &pm, const CompilerOptions &options,
+                                CompilerOutput &output, Pipeline &pipeline,
+                                bool clHasManualPipeline, mlir::ModuleOp moduleOp);
+
+/**
+ * @brief Configure the compiler pass pipeline.
+ *
+ * @param pm
+ * @param options Compiler configuration options.
+ * @param pipeline
+ * @param clHasManualPipeline
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult configurePipeline(mlir::PassManager &pm, const CompilerOptions &options,
+                                      Pipeline &pipeline, bool clHasManualPipeline);
+
+/**
+ * @brief Configure the pass-manager with the given inputs.
+ *
+ * @param pm
+ * @param options Compiler configuration options.
+ * @param output
+ * @param timer
+ * @param timing
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult preparePassManager(mlir::PassManager &pm, const CompilerOptions &options,
+                                       CompilerOutput &output, catalyst::utils::Timer<> &timer,
+                                       mlir::TimingScope &timing);
+
+/**
+ * @brief Buffer and read the given filename contents into a string.
+ *
+ * @param filename
+ * @return std::string
+ */
+std::string readInputFile(const std::string &filename);
+
+/**
+ * @brief Run Enzyme-specific compiler passes.
+ *
+ * @param options Compiler configuration options.
+ * @param llvmModule
+ * @param output
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult runEnzymePasses(const CompilerOptions &options,
+                                    std::shared_ptr<llvm::Module> llvmModule,
+                                    CompilerOutput &output);
+
+/**
+ * @brief Run optimization passes at the -O2 level on the program representation.
+ *
+ * @param options Compiler configuration options.
+ * @param llvmModule
+ * @param output
+ * @return llvm::LogicalResult
+ */
+llvm::LogicalResult runO2LLVMPasses(const CompilerOptions &options,
+                                    std::shared_ptr<llvm::Module> llvmModule,
+                                    CompilerOutput &output);
+
+}; // namespace catalyst::driver
+
+/**
+ * @brief Entry point to the MLIR portion of the compiler.
+ *
+ * @param options
+ * @param output
+ * @param registry
+ * @return mlir::LogicalResult
+ */
 mlir::LogicalResult QuantumDriverMain(const catalyst::driver::CompilerOptions &options,
                                       catalyst::driver::CompilerOutput &output,
                                       mlir::DialectRegistry &registry);
 
+/**
+ * @brief Entry point to the MLIR portion of the compiler using the command-line interface.
+ *
+ * @param argc `main` function argc
+ * @param argv `main` function argv
+ * @return int Return code of execution.
+ */
 int QuantumDriverMainFromCL(int argc, char **argv);
 int QuantumDriverMainFromArgs(const std::string &source, const std::string &workspace,
                               const std::string &moduleName, bool keepIntermediate,
@@ -143,19 +363,3 @@ int QuantumDriverMainFromArgs(const std::string &source, const std::string &work
                               const std::vector<catalyst::driver::Pipeline> &passPipelines,
                               const std::string &checkpointStage,
                               catalyst::driver::CompilerOutput &output);
-
-namespace llvm {
-
-inline raw_ostream &operator<<(raw_ostream &oss, const catalyst::driver::Pipeline &p)
-{
-    oss << "Pipeline('" << p.getName() << "', [";
-    bool first = true;
-    for (const auto &i : p.getPasses()) {
-        oss << (first ? "" : ", ") << i;
-        first = false;
-    }
-    oss << "])";
-    return oss;
-}
-
-}; // namespace llvm
