+++ mlir/lib/Driver/CMakeLists.txt
@@ -55,9 +55,12 @@ set(LIBS
 )
 
 add_mlir_library(CatalystCompilerDriver
+    Main.cpp
     CompilerDriver.cpp
     CatalystLLVMTarget.cpp
+    PassInstrumentation.cpp
     Pipelines.cpp
+    LineUtils.cpp
     HighResolutionOutputStrategy.cpp
 
     LINK_LIBS PRIVATE
