+++ mlir/lib/Driver/CompilerDriver.cpp
@@ -44,7 +44,6 @@
 #include "llvm/Transforms/Coroutines/CoroSplit.h"
 #include "llvm/Transforms/IPO/GlobalDCE.h"
 
-#include "mlir/Bytecode/BytecodeWriter.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/IR/DialectRegistry.h"
 #include "mlir/InitAllDialects.h"
@@ -67,8 +66,11 @@
 #include "Driver/CatalystLLVMTarget.h"
 #include "Driver/CompilerDriver.h"
 #include "Driver/HighResolutionOutputStrategy.h"
+#include "Driver/LineUtils.h"
+#include "Driver/PassInstrumentation.h"
 #include "Driver/Pipelines.h"
 #include "Driver/Support.h"
+#include "Driver/Timer.h"
 #include "Gradient/IR/GradientDialect.h"
 #include "Gradient/IR/GradientInterfaces.h"
 #include "Gradient/IR/GradientOps.h"
@@ -86,126 +88,11 @@
 #include "RegisterAllPasses.h"
 
 #include "Enzyme.h"
-#include "Timer.hpp"
 
 using namespace mlir;
 using namespace catalyst;
 using namespace catalyst::driver;
 namespace cl = llvm::cl;
-
-namespace catalyst::utils {
-
-/**
- * LinesCount : A utility class to count the number of lines of embedded programs
- * in different compilation stages.
- *
- * You can dump the program-size embedded in an `Operation`, `ModuleOp`, or
- * `llvm::Module` using the static methods in this class.
- *
- * To display results, run the driver with the `ENABLE_DIAGNOSTICS=ON` variable.
- * To store results in YAML format, use `DIAGNOSTICS_RESULTS_PATH=/path/to/file.yml`
- * along with `ENABLE_DIAGNOSTICS=ON`.
- */
-class LinesCount {
-  private:
-    inline static void print(const std::string &opStrBuf, const std::string &name)
-    {
-        const auto num_lines = std::count(opStrBuf.cbegin(), opStrBuf.cend(), '\n');
-        if (!name.empty()) {
-            std::cerr << "[DIAGNOSTICS] After " << std::setw(25) << std::left << name;
-        }
-        std::cerr << "\t" << std::fixed << "programsize: " << num_lines << std::fixed << " lines\n";
-    }
-
-    inline static void store(const std::string &opStrBuf, const std::string &name,
-                             const std::filesystem::path &file_path)
-    {
-        const auto num_lines = std::count(opStrBuf.cbegin(), opStrBuf.cend(), '\n');
-
-        const std::string_view key_padding = "          ";
-        const std::string_view val_padding = "              ";
-
-        if (!std::filesystem::exists(file_path)) {
-            std::ofstream ofile(file_path);
-            assert(ofile.is_open() && "Invalid file to store timer results");
-            if (!name.empty()) {
-                ofile << key_padding << "- " << name << ":\n";
-            }
-            ofile << val_padding << "programsize: " << num_lines << "\n";
-            ofile.close();
-            return;
-        }
-        // else
-
-        // Second, update the file
-        std::ofstream ofile(file_path, std::ios::app);
-        assert(ofile.is_open() && "Invalid file to store timer results");
-        if (!name.empty()) {
-            ofile << key_padding << "- " << name << ":\n";
-        }
-        ofile << val_padding << "programsize: " << num_lines << "\n";
-        ofile.close();
-    }
-
-    inline static void dump(const std::string &opStrBuf, const std::string &name = {})
-    {
-        char *file = getenv("DIAGNOSTICS_RESULTS_PATH");
-        if (!file) {
-            print(opStrBuf, name);
-            return;
-        }
-        // else
-        store(opStrBuf, name, std::filesystem::path{file});
-    }
-
-  public:
-    [[nodiscard]] inline static bool is_diagnostics_enabled()
-    {
-        char *value = getenv("ENABLE_DIAGNOSTICS");
-        return value && std::string(value) == "ON";
-    }
-
-    static void Operation(Operation *op, const std::string &name = {})
-    {
-        if (!is_diagnostics_enabled()) {
-            return;
-        }
-
-        std::string opStrBuf;
-        llvm::raw_string_ostream rawStrBef{opStrBuf};
-        rawStrBef << *op;
-
-        dump(opStrBuf, name);
-    }
-
-    static void ModuleOp(const ModuleOp &op, const std::string &name = {})
-    {
-        if (!is_diagnostics_enabled()) {
-            return;
-        }
-
-        std::string modStrBef;
-        llvm::raw_string_ostream rawStrBef{modStrBef};
-        op->print(rawStrBef);
-
-        dump(modStrBef, name);
-    }
-
-    static void Module(const llvm::Module &llvmModule, const std::string &name = {})
-    {
-        if (!is_diagnostics_enabled()) {
-            return;
-        }
-
-        std::string modStrBef;
-        llvm::raw_string_ostream rawStrBef{modStrBef};
-        llvmModule.print(rawStrBef, nullptr);
-
-        dump(modStrBef, name);
-    }
-};
-
-} // namespace catalyst::utils
 
 namespace {
 
@@ -234,107 +121,6 @@ struct CatalystIRPrinterConfig : public PassManager::IRPrinterConfig {
     }
 };
 
-struct CatalystPassInstrumentation : public PassInstrumentation {
-    const CompilerOptions &options;
-    CompilerOutput &output;
-    catalyst::utils::Timer &timer;
-
-    // Store fingerprints before each pass to detect changes
-    DenseMap<Pass *, std::optional<OperationFingerPrint>> beforePassFingerprints;
-
-    CatalystPassInstrumentation(const CompilerOptions &options, CompilerOutput &output,
-                                catalyst::utils::Timer &timer)
-        : options(options), output(output), timer(timer)
-    {
-    }
-
-    void runBeforePass(Pass *pass, Operation *operation) override
-    {
-        if (this->options.verbosity >= Verbosity::Debug && !this->timer.is_active()) {
-            this->timer.start();
-        }
-        if (this->options.keepIntermediate == SaveTemps::AfterPassChanged) {
-            this->beforePassFingerprints[pass] = OperationFingerPrint(operation);
-        }
-    }
-
-    void runAfterPass(Pass *pass, Operation *operation) override
-    {
-        // Handle verbosity logging
-        if (this->options.verbosity >= Verbosity::Debug) {
-            auto pipelineName = pass->getName();
-            this->timer.dump(pipelineName.str(), /*add_endl */ false);
-            catalyst::utils::LinesCount::Operation(operation);
-        }
-
-        bool shouldDump = false;
-
-        if (this->options.keepIntermediate == SaveTemps::AfterPass) {
-            shouldDump = true;
-        }
-        else if (this->options.keepIntermediate == SaveTemps::AfterPassChanged) {
-            // If change detection is enabled, only dump if IR is changed
-            auto it = this->beforePassFingerprints.find(pass);
-            if (it != this->beforePassFingerprints.end() && it->second.has_value()) {
-                OperationFingerPrint afterFingerprint(operation);
-                shouldDump = (afterFingerprint != it->second.value());
-                this->beforePassFingerprints.erase(it);
-            }
-            else {
-                // Fingerprint not found: default to dumping to be safe
-                shouldDump = true;
-            }
-        }
-
-        if (shouldDump) {
-            this->dumpIRAfterPass(pass, operation);
-        }
-    }
-
-    void runAfterPassFailed(Pass *pass, Operation *operation) override
-    {
-        // Always dump on failure for debugging
-        this->options.diagnosticStream << "While processing '" << pass->getName().str()
-                                       << "' pass ";
-        std::string tmp;
-        llvm::raw_string_ostream s{tmp};
-        s << *operation;
-        if (this->options.keepIntermediate) {
-            dumpToFile(this->options,
-                       this->output.nextPipelineSummaryFilename(pass->getName().str() + "_FAILED"),
-                       tmp);
-        }
-
-        // Clean up fingerprint if present
-        if (this->options.keepIntermediate == SaveTemps::AfterPassChanged) {
-            this->beforePassFingerprints.erase(pass);
-        }
-    }
-
-  private:
-    void dumpIRAfterPass(Pass *pass, Operation *op)
-    {
-        auto pipelineName = pass->getName();
-
-        // Save IR after pass
-        std::string tmp;
-        llvm::raw_string_ostream s{tmp};
-        if (this->options.dumpModuleScope) {
-            mlir::ModuleOp mod = isa<mlir::ModuleOp>(op) ? cast<mlir::ModuleOp>(op)
-                                                         : op->getParentOfType<mlir::ModuleOp>();
-            s << mod;
-        }
-        else {
-            s << *op;
-        }
-        std::string fileName = pipelineName.str();
-        if (auto funcOp = dyn_cast<mlir::func::FuncOp>(op)) {
-            fileName += std::string("_") + funcOp.getName().str();
-        }
-        dumpToFile(this->options, this->output.nextPassDumpFilename(fileName), tmp);
-    }
-};
-
 } // namespace
 
 /// The upstream MLIR Test dialect does not have a header we can include
@@ -344,7 +130,7 @@ namespace test {
 void registerTestDialect(mlir::DialectRegistry &);
 } // namespace test
 
-namespace {
+namespace catalyst::driver {
 /// Parse an MLIR module given in textual ASM representation. Any errors during parsing will be
 /// output to diagnosticStream.
 OwningOpRef<ModuleOp> parseMLIRSource(MLIRContext *ctx, const llvm::SourceMgr &sourceMgr)
@@ -400,11 +186,11 @@ void registerAllCatalystDialects(DialectRegistry &registry)
     registry.insert<qecl::QecLogicalDialect>();
     registry.insert<qecp::QecPhysicalDialect>();
 }
-} // namespace
+} // namespace catalyst::driver
 
 // Determines if the compilation stage should be executed if a checkpointStage is given
-bool shouldRunStage(const CompilerOptions &options, CompilerOutput &output,
-                    const std::string &stageName)
+bool catalyst::driver::shouldRunStage(const CompilerOptions &options, CompilerOutput &output,
+                                      const std::string &stageName)
 {
     if (options.checkpointStage.empty()) {
         return true;
@@ -416,10 +202,11 @@ bool shouldRunStage(const CompilerOptions &options, CompilerOutput &output,
     return true;
 }
 
-LogicalResult runCoroLLVMPasses(const CompilerOptions &options,
-                                std::shared_ptr<llvm::Module> llvmModule, CompilerOutput &output)
+llvm::LogicalResult catalyst::driver::runCoroLLVMPasses(const CompilerOptions &options,
+                                                        std::shared_ptr<llvm::Module> llvmModule,
+                                                        CompilerOutput &output)
 {
-    if (!shouldRunStage(options, output, "CoroOpt")) {
+    if (!catalyst::driver::shouldRunStage(options, output, "CoroOpt")) {
         return success();
     }
 
@@ -460,13 +247,14 @@ LogicalResult runCoroLLVMPasses(const CompilerOptions &options,
     return success();
 }
 
-LogicalResult runO2LLVMPasses(const CompilerOptions &options,
-                              std::shared_ptr<llvm::Module> llvmModule, CompilerOutput &output)
+llvm::LogicalResult catalyst::driver::runO2LLVMPasses(const CompilerOptions &options,
+                                                      std::shared_ptr<llvm::Module> llvmModule,
+                                                      CompilerOutput &output)
 {
     // opt -O2
     // As seen here:
     // https://llvm.org/docs/NewPassManager.html#just-tell-me-how-to-run-the-default-optimization-pipeline-with-the-new-pass-manager
-    if (!shouldRunStage(options, output, "O2Opt")) {
+    if (!catalyst::driver::shouldRunStage(options, output, "O2Opt")) {
         return success();
     }
 
@@ -513,10 +301,11 @@ LogicalResult runO2LLVMPasses(const CompilerOptions &options,
     return success();
 }
 
-LogicalResult runEnzymePasses(const CompilerOptions &options,
-                              std::shared_ptr<llvm::Module> llvmModule, CompilerOutput &output)
+llvm::LogicalResult catalyst::driver::runEnzymePasses(const CompilerOptions &options,
+                                                      std::shared_ptr<llvm::Module> llvmModule,
+                                                      CompilerOutput &output)
 {
-    if (!shouldRunStage(options, output, "Enzyme")) {
+    if (!catalyst::driver::shouldRunStage(options, output, "Enzyme")) {
         return success();
     }
 
@@ -562,7 +351,7 @@ LogicalResult runEnzymePasses(const CompilerOptions &options,
     return success();
 }
 
-std::string readInputFile(const std::string &filename)
+std::string catalyst::driver::readInputFile(const std::string &filename)
 {
     if (filename == "-") {
         if (llvm::errs().is_displayed()) {
@@ -582,9 +371,10 @@ std::string readInputFile(const std::string &filename)
     return buffer.str();
 }
 
-LogicalResult preparePassManager(PassManager &pm, const CompilerOptions &options,
-                                 CompilerOutput &output, catalyst::utils::Timer &timer,
-                                 TimingScope &timing)
+[[nodiscard]] llvm::LogicalResult
+catalyst::driver::preparePassManager(PassManager &pm, const CompilerOptions &options,
+                                     CompilerOutput &output, catalyst::utils::Timer<> &timer,
+                                     TimingScope &timing)
 {
     MlirOptMainConfig config = MlirOptMainConfig::createFromCLOptions();
     pm.enableVerifier(config.shouldVerifyPasses());
@@ -599,8 +389,9 @@ LogicalResult preparePassManager(PassManager &pm, const CompilerOptions &options
     return success();
 }
 
-LogicalResult configurePipeline(PassManager &pm, const CompilerOptions &options, Pipeline &pipeline,
-                                bool clHasManualPipeline)
+[[nodiscard]] llvm::LogicalResult
+catalyst::driver::configurePipeline(PassManager &pm, const CompilerOptions &options,
+                                    Pipeline &pipeline, bool clHasManualPipeline)
 {
     pm.clear();
     if (!clHasManualPipeline && failed(pipeline.addPipeline(pm))) {
@@ -618,16 +409,18 @@ LogicalResult configurePipeline(PassManager &pm, const CompilerOptions &options,
     return success();
 }
 
-LogicalResult runPipeline(PassManager &pm, const CompilerOptions &options, CompilerOutput &output,
-                          Pipeline &pipeline, bool clHasManualPipeline, ModuleOp moduleOp)
+llvm::LogicalResult catalyst::driver::runPipeline(PassManager &pm, const CompilerOptions &options,
+                                                  CompilerOutput &output, Pipeline &pipeline,
+                                                  bool clHasManualPipeline, ModuleOp moduleOp)
 {
-    if (!shouldRunStage(options, output, pipeline.getName()) || pipeline.getPasses().size() == 0) {
+    if (!catalyst::driver::shouldRunStage(options, output, pipeline.getName()) ||
+        pipeline.getPasses().size() == 0) {
         return success();
     }
 
     output.setStage(pipeline.getName());
 
-    if (failed(configurePipeline(pm, options, pipeline, clHasManualPipeline))) {
+    if (failed(catalyst::driver::configurePipeline(pm, options, pipeline, clHasManualPipeline))) {
         llvm::errs() << "Failed to run pipeline: " << pipeline.getName() << "\n";
         return failure();
     }
@@ -644,14 +437,16 @@ LogicalResult runPipeline(PassManager &pm, const CompilerOptions &options, Compi
     return success();
 }
 
-LogicalResult runLowering(const CompilerOptions &options, MLIRContext *ctx, ModuleOp moduleOp,
-                          CompilerOutput &output, TimingScope &timing)
+[[nodiscard]] llvm::LogicalResult catalyst::driver::runLowering(const CompilerOptions &options,
+                                                                MLIRContext *ctx, ModuleOp moduleOp,
+                                                                CompilerOutput &output,
+                                                                TimingScope &timing)
 
 {
-    catalyst::utils::Timer timer{};
+    catalyst::utils::Timer<> timer{};
 
     auto pm = PassManager::on<ModuleOp>(ctx, PassManager::Nesting::Implicit);
-    if (failed(preparePassManager(pm, options, output, timer, timing))) {
+    if (failed(catalyst::driver::preparePassManager(pm, options, output, timer, timing))) {
         llvm::errs() << "Failed to setup pass manager\n";
         return failure();
     }
@@ -677,17 +472,19 @@ LogicalResult runLowering(const CompilerOptions &options, MLIRContext *ctx, Modu
     std::vector<Pipeline> UserPipeline =
         clHasManualPipeline ? options.pipelinesCfg : getDefaultPipeline();
     for (auto &pipeline : UserPipeline) {
-        if (failed(catalyst::utils::Timer::timer(runPipeline, pipeline.getName(),
-                                                 /* add_endl */ false, pm, options, output,
-                                                 pipeline, clHasManualPipeline, moduleOp))) {
+        if (failed(catalyst::utils::Timer<>::timer(catalyst::driver::runPipeline,
+                                                   pipeline.getName(),
+                                                   /* add_endl */ false, pm, options, output,
+                                                   pipeline, clHasManualPipeline, moduleOp))) {
             return failure();
         }
-        catalyst::utils::LinesCount::ModuleOp(moduleOp);
+        catalyst::utils::LinesCount::call(moduleOp);
     }
     return success();
 }
 
-LogicalResult verifyInputType(const CompilerOptions &options, InputType inType)
+llvm::LogicalResult catalyst::driver::verifyInputType(const CompilerOptions &options,
+                                                      InputType inType)
 {
     if (inType == InputType::OTHER) {
         CO_MSG(options, Verbosity::Urgent, "Wrong or unsupported input\n");
@@ -704,246 +501,7 @@ LogicalResult verifyInputType(const CompilerOptions &options, InputType inType)
     return success();
 }
 
-LogicalResult QuantumDriverMain(const CompilerOptions &options, CompilerOutput &output,
-                                DialectRegistry &registry)
-{
-    using timer = catalyst::utils::Timer;
-
-    OpPrintingFlags opPrintingFlags{};
-    if (options.useNameLocAsPrefix) {
-        opPrintingFlags.printNameLocAsPrefix();
-    }
-
-    MLIRContext ctx(registry);
-    ctx.printOpOnDiagnostic(true);
-    ctx.printStackTraceOnDiagnostic(options.verbosity >= Verbosity::Debug);
-    // TODO: FIXME:
-    // Let's try to enable multithreading. Do not forget to protect the printing.
-    ctx.disableMultithreading();
-    // The transform dialect doesn't appear to load dependent dialects
-    // fpr named passes.
-    ctx.loadAllAvailableDialects();
-
-    ScopedDiagnosticHandler scopedHandler(
-        &ctx, [&](Diagnostic &diag) { diag.print(options.diagnosticStream); });
-
-    llvm::LLVMContext llvmContext;
-    std::shared_ptr<llvm::Module> llvmModule;
-
-    llvm::raw_string_ostream outIRStream(output.outIR);
-
-    auto moduleBuffer = llvm::MemoryBuffer::getMemBufferCopy(options.source, options.moduleName);
-    auto sourceMgr = std::make_shared<llvm::SourceMgr>();
-    sourceMgr->AddNewSourceBuffer(std::move(moduleBuffer), SMLoc());
-    SourceMgrDiagnosticHandler sourceMgrHandler(*sourceMgr, &ctx, options.diagnosticStream);
-
-    DefaultTimingManager tm;
-    applyDefaultTimingManagerCLOptions(tm);
-    if (tm.isEnabled()) {
-        auto strategy = std::make_unique<HighResolutionOutputStrategy>(llvm::errs());
-        tm.setOutput(std::move(strategy));
-    }
-    TimingScope timing = tm.getRootScope();
-
-    TimingScope parserTiming = timing.nest("Parser");
-    OwningOpRef<ModuleOp> mlirModule = timer::timer(parseMLIRSource, "parseMLIRSource",
-                                                    /* add_endl */ false, &ctx, *sourceMgr);
-
-    enum InputType inType = InputType::OTHER;
-    if (mlirModule) {
-        inType = InputType::MLIR;
-        catalyst::utils::LinesCount::ModuleOp(*mlirModule);
-        output.isCheckpointFound = options.checkpointStage == "mlir";
-
-        if (options.verbosity == Verbosity::All) {
-            llvm::outs() << "MLIR parsing successful" << "\n";
-        }
-    }
-    else {
-        llvm::SMDiagnostic err;
-        llvmModule = timer::timer(parseLLVMSource, "parseLLVMSource", false, llvmContext,
-                                  options.source, options.moduleName, err);
-
-        if (!llvmModule) {
-            err.print(options.moduleName.data(), options.diagnosticStream);
-            CO_MSG(options, Verbosity::Urgent, "Failed to parse module as LLVM or MLIR source\n");
-            return failure();
-        }
-        inType = InputType::LLVMIR;
-        output.isCheckpointFound = options.checkpointStage == "LLVMIRTranslation";
-        catalyst::utils::LinesCount::Module(*llvmModule);
-
-        if (options.verbosity == Verbosity::All) {
-            llvm::outs() << "LLVMIR parsing successful" << "\n";
-        }
-    }
-    if (failed(verifyInputType(options, inType))) {
-        return failure();
-    }
-    parserTiming.stop();
-
-    bool runAll = (options.loweringAction == Action::All);
-    bool runOpt = (options.loweringAction == Action::OPT) || runAll;
-    bool runTranslate = (options.loweringAction == Action::Translate) || runAll;
-    bool runLLC = (options.loweringAction == Action::LLC) || runAll;
-
-    if (runOpt && (inType == InputType::MLIR)) {
-        TimingScope optTiming = timing.nest("Optimization");
-        if (failed(runLowering(options, &ctx, *mlirModule, output, optTiming))) {
-            CO_MSG(options, Verbosity::Urgent, "Failed to lower MLIR module\n");
-            return failure();
-        }
-        output.outIR.clear();
-        if (options.keepIntermediate) {
-            mlirModule->print(outIRStream, opPrintingFlags);
-        }
-        optTiming.stop();
-
-        if (options.verbosity == Verbosity::All) {
-            llvm::outs() << "quantum-opt transformations successful" << "\n";
-        }
-    }
-
-    if (runTranslate && (inType == InputType::MLIR)) {
-        TimingScope translateTiming = timing.nest("Translate");
-        llvmModule =
-            timer::timer(translateModuleToLLVMIR, "translateModuleToLLVMIR",
-                         /* add_endl */ false, *mlirModule, llvmContext, "LLVMDialectModule",
-                         /* disableVerification */ true);
-        if (!llvmModule) {
-            CO_MSG(options, Verbosity::Urgent, "Failed to translate LLVM module\n");
-            return failure();
-        }
-
-        inType = InputType::LLVMIR;
-        catalyst::utils::LinesCount::Module(*llvmModule);
-
-        if (options.keepIntermediate) {
-            output.setStage("LLVMIRTranslation");
-            std::string tmp;
-            llvm::raw_string_ostream rawStringOstream{tmp};
-            llvmModule->print(rawStringOstream, nullptr);
-            auto outFile = output.nextPipelineSummaryFilename("LLVMIRTranslation", ".ll");
-            dumpToFile(options, outFile, tmp);
-        }
-        output.outIR.clear();
-        if (options.keepIntermediate) {
-            outIRStream << *llvmModule;
-        }
-        translateTiming.stop();
-
-        if (options.verbosity == Verbosity::All) {
-            llvm::outs() << "MLIR->LLVMIR translation successful" << "\n";
-        }
-    }
-
-    if (runLLC && (inType == InputType::LLVMIR)) {
-        TimingScope llcTiming = timing.nest("llc");
-        // Set data layout before LLVM passes or the default one is used.
-        llvm::Triple targetTriple(llvm::sys::getDefaultTargetTriple());
-
-        llvm::InitializeAllTargetInfos();
-        llvm::InitializeAllTargets();
-        llvm::InitializeAllTargetMCs();
-        llvm::InitializeAllAsmParsers();
-        llvm::InitializeAllAsmPrinters();
-
-        std::string err;
-        auto target = llvm::TargetRegistry::lookupTarget(targetTriple, err);
-        llvm::TargetOptions opt;
-        const char *cpu = "generic";
-        const char *features = "";
-        auto targetMachine =
-            target->createTargetMachine(targetTriple, cpu, features, opt, llvm::Reloc::Model::PIC_);
-        targetMachine->setOptLevel(llvm::CodeGenOptLevel::None);
-        llvmModule->setDataLayout(targetMachine->createDataLayout());
-        llvmModule->setTargetTriple(targetTriple);
-
-        if (options.asyncQnodes) {
-            TimingScope coroLLVMPassesTiming = llcTiming.nest("LLVM coroutine passes");
-            if (failed(timer::timer(runCoroLLVMPasses, "runCoroLLVMPasses",
-                                    /* add_endl */ false, options, llvmModule, output))) {
-                return failure();
-            }
-            coroLLVMPassesTiming.stop();
-            catalyst::utils::LinesCount::Module(*llvmModule.get());
-        }
-
-        bool enzymeRun = containsGradients(*llvmModule);
-        if (enzymeRun) {
-            TimingScope o2PassesTiming = llcTiming.nest("LLVM O2 passes");
-            if (failed(timer::timer(runO2LLVMPasses, "runO2LLVMPasses", /* add_endl */ false,
-                                    options, llvmModule, output))) {
-                return failure();
-            }
-            o2PassesTiming.stop();
-            catalyst::utils::LinesCount::Module(*llvmModule.get());
-
-            TimingScope enzymePassesTiming = llcTiming.nest("Enzyme passes");
-            if (failed(timer::timer(runEnzymePasses, "runEnzymePasses", /* add_endl */ false,
-                                    options, llvmModule, output))) {
-                return failure();
-            }
-            enzymePassesTiming.stop();
-            catalyst::utils::LinesCount::Module(*llvmModule.get());
-        }
-
-        std::string errorMessage;
-        auto outfile = openOutputFile(output.outputFilename, &errorMessage);
-        if (output.outputFilename == "-" && llvmModule) {
-            // Do not generate file if outputting to stdout.
-            outfile->os() << *llvmModule;
-            outfile->keep();
-            // early exit
-            return success();
-        }
-
-        TimingScope outputTiming = llcTiming.nest("compileObject");
-        output.outIR.clear();
-        if (options.keepIntermediate) {
-            outIRStream << *llvmModule;
-        }
-
-        if (failed(timer::timer(compileObjectFile, "compileObjFile", /* add_endl */ true, options,
-                                llvmModule, targetMachine, options.getObjectFile()))) {
-            return failure();
-        }
-        outputTiming.stop();
-        llcTiming.stop();
-
-        if (options.verbosity == Verbosity::All) {
-            llvm::outs() << "LLC object code generation successful" << "\n";
-        }
-    }
-
-    std::string errorMessage;
-    auto outfile = openOutputFile(output.outputFilename, &errorMessage);
-    if (!outfile) {
-        llvm::errs() << errorMessage << "\n";
-        return failure();
-    }
-    else if (output.outputFilename == "-" && llvmModule) {
-        // already handled
-    }
-    else if (output.outputFilename == "-" && mlirModule) {
-        if (options.shouldEmitBytecode) {
-            return mlir::writeBytecodeToFile(mlirModule.get(), outfile->os());
-        }
-        else {
-            mlirModule->print(outfile->os(), opPrintingFlags);
-            outfile->keep();
-        }
-    }
-
-    if (options.keepIntermediate and output.outputFilename != "-") {
-        outfile->os() << output.outIR;
-        outfile->keep();
-    }
-
-    return success();
-}
-
-size_t findMatchingClosingParen(llvm::StringRef str, size_t openParenPos)
+size_t catalyst::driver::findMatchingClosingParen(llvm::StringRef str, size_t openParenPos)
 {
     int parenCount = 1;
     for (size_t pos = openParenPos + 1; pos < str.size(); pos++) {
@@ -960,7 +518,8 @@ size_t findMatchingClosingParen(llvm::StringRef str, size_t openParenPos)
     return llvm::StringRef::npos;
 }
 
-std::vector<Pipeline> parsePipelines(const cl::list<std::string> &catalystPipeline)
+std::vector<Pipeline>
+catalyst::driver::parsePipelines(const cl::list<std::string> &catalystPipeline)
 {
     std::vector<Pipeline> allPipelines;
     for (const auto &pipelineStr : catalystPipeline) {
@@ -971,7 +530,8 @@ std::vector<Pipeline> parsePipelines(const cl::list<std::string> &catalystPipeli
         }
 
         size_t openParenPos = pipelineRef.find('(');
-        size_t closeParenPos = findMatchingClosingParen(pipelineRef, openParenPos);
+        size_t closeParenPos =
+            catalyst::driver::findMatchingClosingParen(pipelineRef, openParenPos);
 
         if (openParenPos == llvm::StringRef::npos || closeParenPos == llvm::StringRef::npos) {
             llvm::errs() << "Error: Invalid pipeline format: " << pipelineStr << "\n";
@@ -997,136 +557,31 @@ std::vector<Pipeline> parsePipelines(const cl::list<std::string> &catalystPipeli
     return allPipelines;
 }
 
-int QuantumDriverMainFromCL(int argc, char **argv)
+std::string CompilerOptions::getObjectFile() const
 {
-    // Command-line options
+    using path = std::filesystem::path;
+    return path(workspace.str()) / path(moduleName.str() + ".o");
+}
 
-    // ATTENTION
-    // ---------
-    // Any modifications made to the command-line interface should be documented in
-    // doc/catalyst-cli/catalyst-cli.rst
-    cl::OptionCategory CatalystCat("Catalyst CLI Options", "");
-    cl::opt<std::string> WorkspaceDir("workspace", cl::desc("Workspace directory"), cl::init("."),
-                                      cl::cat(CatalystCat));
-    cl::opt<std::string> ModuleName("module-name", cl::desc("Module name"),
-                                    cl::init("catalyst_module"), cl::cat(CatalystCat));
+std::string CompilerOutput::nextPassDumpFilename(const std::string &pipelineName,
+                                                 const std::string &ext)
+{
+    return std::filesystem::path(currentStage) /
+           std::filesystem::path(std::to_string(this->passCounter++) + "_" + pipelineName)
+               .replace_extension(ext);
+}
 
-    cl::opt<enum SaveTemps> SaveAfterEach(
-        "save-ir-after-each", cl::desc("Keep intermediate files after each pass or pipeline"),
-        cl::values(clEnumValN(SaveTemps::AfterPassChanged, "changed",
-                              "Save IR after each pass (only if changed)")),
-        cl::values(clEnumValN(SaveTemps::AfterPass, "pass",
-                              "Save IR after each pass (even if unchanged)")),
-        cl::values(clEnumValN(SaveTemps::AfterPipeline, "pipeline", "Save IR after each pipeline")),
-        cl::init(SaveTemps::None), cl::cat(CatalystCat));
-    cl::opt<bool> KeepIntermediate(
-        "keep-intermediate", cl::desc("Keep intermediate files"), cl::init(false),
-        cl::callback([&](const bool &) { SaveAfterEach.setValue(SaveTemps::AfterPipeline); }),
-        cl::cat(CatalystCat));
-    cl::opt<bool> UseNameLocAsPrefix("use-nameloc-as-prefix",
-                                     cl::desc("Use name location as prefix"), cl::init(false),
-                                     cl::cat(CatalystCat));
-    cl::opt<bool> AsyncQNodes("async-qnodes", cl::desc("Enable asynchronous QNodes"),
-                              cl::init(false), cl::cat(CatalystCat));
-    cl::opt<bool> Verbose("verbose", cl::desc("Set verbose"), cl::init(false),
-                          cl::cat(CatalystCat));
-    cl::list<std::string> CatalystPipeline(
-        "catalyst-pipeline", cl::desc("Catalyst Compiler pass pipelines"), cl::ZeroOrMore,
-        cl::CommaSeparated, cl::cat(CatalystCat));
-    cl::opt<std::string> CheckpointStage("checkpoint-stage", cl::desc("Checkpoint stage"),
-                                         cl::init(""), cl::cat(CatalystCat));
-    cl::opt<enum Action> LoweringAction(
-        "tool", cl::desc("Select the tool to isolate"),
-        cl::values(clEnumValN(Action::OPT, "opt", "run quantum-opt on the MLIR input")),
-        cl::values(clEnumValN(Action::Translate, "translate",
-                              "run mlir-translate on the MLIR LLVM dialect")),
-        cl::values(clEnumValN(Action::LLC, "llc", "run llc on the llvm IR input")),
-        cl::values(clEnumValN(Action::All, "all",
-                              "run quantum-opt, mlir-translate, and llc on the MLIR input")),
-        cl::init(Action::All), cl::cat(CatalystCat));
-    cl::opt<bool> DumpPassPipeline("dump-catalyst-pipeline",
-                                   cl::desc("Print the pipeline that will be run"), cl::init(false),
-                                   cl::cat(CatalystCat));
-    cl::opt<bool> DumpModuleScope("dump-module-scope",
-                                  cl::desc("Print the whole module in intermediate files"),
-                                  cl::init(true), cl::cat(CatalystCat));
+std::string CompilerOutput::nextPipelineSummaryFilename(const std::string &pipelineName,
+                                                        const std::string &ext)
+{
+    return std::filesystem::path(std::to_string(this->globalPipelineCounter) + "_After" +
+                                 pipelineName)
+        .replace_extension(ext);
+}
 
-    // Create dialect registry
-    DialectRegistry registry;
-    mlir::registerAllPasses();
-    catalyst::registerAllPasses();
-    registerAllCatalystPipelines();
-    mlirRegisterAllStablehloPasses();
-    mlir::stablehlo::registerOptimizationPasses();
-    registerAllCatalystDialects(registry);
-    registerLLVMTranslations(registry);
-
-    // Register bufferization interfaces
-    catalyst::registerBufferizableOpInterfaceExternalModels(registry);
-    catalyst::gradient::registerBufferizableOpInterfaceExternalModels(registry);
-    catalyst::quantum::registerBufferizableOpInterfaceExternalModels(registry);
-
-    // Register and parse command line options.
-    std::string inputFilename, outputFilename;
-    std::string helpStr = "Catalyst Command Line Interface options. \n"
-                          "Below, there is a complete list of options for the Catalyst CLI tool"
-                          "In the first section, you can find the options that are used to"
-                          "configure the Catalyst compiler. Next, you can find the options"
-                          "specific to the mlir-opt tool.\n";
-    std::tie(inputFilename, outputFilename) =
-        registerAndParseCLIOptions(argc, argv, helpStr, registry);
-    llvm::InitLLVM y(argc, argv);
-    MlirOptMainConfig config = MlirOptMainConfig::createFromCLOptions();
-
-    if (config.shouldShowDialects()) {
-        llvm::outs() << "Available Dialects: ";
-        interleave(registry.getDialectNames(), llvm::outs(), ",");
-        llvm::outs() << "\n";
-        return 0;
-    }
-
-    if (config.shouldListPasses()) {
-        mlir::printRegisteredPasses();
-        return 0;
-    }
-
-    std::unique_ptr<CompilerOutput> output(new CompilerOutput());
-    assert(output);
-    output->outputFilename = outputFilename;
-    llvm::raw_string_ostream errStream{output->diagnosticMessages};
-
-    // Read the input IR file
-    std::string source = readInputFile(inputFilename);
-    if (source.empty()) {
-        llvm::errs() << "Error: Unable to read input file: " << inputFilename << "\n";
-        return 1;
-    }
-
-    CompilerOptions options{.source = source,
-                            .workspace = WorkspaceDir,
-                            .moduleName = ModuleName,
-                            .diagnosticStream = errStream,
-                            .keepIntermediate = SaveAfterEach,
-                            .dumpModuleScope = DumpModuleScope,
-                            .useNameLocAsPrefix = UseNameLocAsPrefix,
-                            .asyncQnodes = AsyncQNodes,
-                            .verbosity = Verbose ? Verbosity::All : Verbosity::Urgent,
-                            .pipelinesCfg = parsePipelines(CatalystPipeline),
-                            .checkpointStage = CheckpointStage,
-                            .loweringAction = LoweringAction,
-                            .dumpPassPipeline = DumpPassPipeline,
-                            .shouldEmitBytecode = config.shouldEmitBytecode()};
-
-    mlir::LogicalResult result = QuantumDriverMain(options, *output, registry);
-
-    errStream.flush();
-
-    if (mlir::failed(result)) {
-        llvm::errs() << "Compilation failed:\n" << output->diagnosticMessages << "\n";
-        return 1;
-    }
-
-    if (Verbose)
-        llvm::outs() << "Compilation successful:\n" << output->diagnosticMessages << "\n";
-    return 0;
+void CompilerOutput::setStage(const std::string &stageName)
+{
+    ++globalPipelineCounter;
+    currentStage = std::to_string(globalPipelineCounter) + "_" + stageName;
+    passCounter = 1;
 }
