+++ runtime/lib/capi/RuntimeCAPI.cpp
@@ -24,11 +24,11 @@
 
 #include "mlir/ExecutionEngine/CRunnerUtils.h"
 
+#include "Driver/Timer.h"
 #include "Exception.hpp"
 #include "ExecutionContext.hpp"
 #include "MemRefUtils.hpp"
 #include "QuantumDevice.hpp"
-#include "Timer.hpp"
 #include "Types.h"
 
 #include "RuntimeCAPI.h"
@@ -122,7 +122,7 @@ static void autoQubitManagementAllocate(std::vector<QubitIdType> *qubit_vector_p
 extern "C" {
 
 using namespace Catalyst::Runtime;
-using timer = catalyst::utils::Timer;
+using timer = catalyst::utils::Timer<>; // default timing ratio of std::milli
 
 void __catalyst_inactive_callback(int64_t identifier, int64_t argc, int64_t retc, ...)
 {
